pyproject.toml
[tool.poetry]
name = "h3daemon"
version = "0.1.0"
description = "Run HMMER daemon on containers"
authors = ["Danilo Horta <danilo.horta@pm.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
podman = "^4.3.0"
typer = {extras = ["all"], version = "^0.7.0"}
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
h3d = 'h3daemon.cli:app'
